+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/__init__.py
@@ -1,26 +1,40 @@
 #
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
-from dagster import Definitions, ScheduleDefinition, load_assets_from_modules
+from dagster import Definitions, ScheduleDefinition, EnvVar, load_assets_from_modules
+from dagster_slack import SlackResource
 
 from orchestrator.resources.gcp import gcp_gcs_client, gcs_directory_blobs, gcs_file_blob, gcs_file_manager
-from orchestrator.resources.github import github_client, github_connector_repo, github_connectors_directory, github_workflow_runs
+from orchestrator.resources.github import (
+    github_client,
+    github_connector_repo,
+    github_connectors_directory,
+    github_workflow_runs,
+    github_connectors_metadata_files,
+)
 
 from orchestrator.assets import (
     connector_test_report,
     github,
     specs_secrets_mask,
-    spec_cache,
     registry,
     registry_report,
     registry_entry,
     metadata,
 )
 
-from orchestrator.jobs.registry import generate_registry_reports, generate_oss_registry, generate_cloud_registry, generate_registry_entry
+from orchestrator.jobs.registry import (
+    generate_registry_reports,
+    generate_oss_registry,
+    generate_cloud_registry,
+    generate_registry_entry,
+    add_new_metadata_partitions,
+)
 from orchestrator.jobs.connector_test_report import generate_nightly_reports, generate_connector_test_summary_reports
+from orchestrator.jobs.metadata import generate_stale_gcs_latest_metadata_file
 from orchestrator.sensors.registry import registry_updated_sensor
-from orchestrator.sensors.gcs import new_gcs_blobs_sensor, new_gcs_blobs_partition_sensor
+from orchestrator.sensors.gcs import new_gcs_blobs_sensor
+from orchestrator.logging.sentry import setup_dagster_sentry
 
 from orchestrator.config import (
     REPORT_FOLDER,
@@ -33,6 +47,7 @@ from orchestrator.config import (
     NIGHTLY_GHA_WORKFLOW_ID,
     CI_TEST_REPORT_PREFIX,
     CI_MASTER_TEST_OUTPUT_REGEX,
+    HIGH_QUEUE_PRIORITY,
 )
 from metadata_service.constants import METADATA_FILE_NAME, METADATA_FOLDER
 
@@ -40,7 +55,6 @@ ASSETS = load_assets_from_modules(
     [
         github,
         specs_secrets_mask,
-        spec_cache,
         metadata,
         registry,
         registry_report,
@@ -49,11 +63,15 @@ ASSETS = load_assets_from_modules(
     ]
 )
 
+SLACK_RESOURCE_TREE = {
+    "slack": SlackResource(token=EnvVar("SLACK_TOKEN")),
+}
 
-RESOURCES = {
+GITHUB_RESOURCE_TREE = {
     "github_client": github_client.configured({"github_token": {"env": "GITHUB_METADATA_SERVICE_TOKEN"}}),
     "github_connector_repo": github_connector_repo.configured({"connector_repo_name": CONNECTOR_REPO_NAME}),
     "github_connectors_directory": github_connectors_directory.configured({"connectors_path": CONNECTORS_PATH}),
+    "github_connectors_metadata_files": github_connectors_metadata_files.configured({"connectors_path": CONNECTORS_PATH}),
     "github_connector_nightly_workflow_successes": github_workflow_runs.configured(
         {
             "workflow_id": NIGHTLY_GHA_WORKFLOW_ID,
@@ -61,6 +79,9 @@ RESOURCES = {
             "status": "success",
         }
     ),
+}
+
+GCS_RESOURCE_TREE = {
     "gcp_gcs_client": gcp_gcs_client.configured(
         {
             "gcp_gcs_cred_string": {"env": "GCS_CREDENTIALS"},
@@ -69,24 +90,45 @@ RESOURCES = {
     "registry_directory_manager": gcs_file_manager.configured({"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": REGISTRIES_FOLDER}),
     "registry_report_directory_manager": gcs_file_manager.configured({"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": REPORT_FOLDER}),
     "root_metadata_directory_manager": gcs_file_manager.configured({"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": ""}),
+}
+
+METADATA_RESOURCE_TREE = {
+    **SLACK_RESOURCE_TREE,
+    **GCS_RESOURCE_TREE,
     "all_metadata_file_blobs": gcs_directory_blobs.configured(
         {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": METADATA_FOLDER, "match_regex": f".*/{METADATA_FILE_NAME}$"}
     ),
     "latest_metadata_file_blobs": gcs_directory_blobs.configured(
         {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": METADATA_FOLDER, "match_regex": f".*latest/{METADATA_FILE_NAME}$"}
     ),
-    "latest_cloud_registry_entries_file_blobs": gcs_directory_blobs.configured(
-        {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": METADATA_FOLDER, "match_regex": f".*latest/cloud.json$"}
-    ),
-    "latest_oss_registry_entries_file_blobs": gcs_directory_blobs.configured(
-        {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": METADATA_FOLDER, "match_regex": f".*latest/oss.json$"}
-    ),
+}
+
+REGISTRY_RESOURCE_TREE = {
+    **SLACK_RESOURCE_TREE,
+    **GCS_RESOURCE_TREE,
     "latest_oss_registry_gcs_blob": gcs_file_blob.configured(
         {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": REGISTRIES_FOLDER, "gcs_filename": "oss_registry.json"}
     ),
     "latest_cloud_registry_gcs_blob": gcs_file_blob.configured(
         {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": REGISTRIES_FOLDER, "gcs_filename": "cloud_registry.json"}
     ),
+}
+
+REGISTRY_ENTRY_RESOURCE_TREE = {
+    **SLACK_RESOURCE_TREE,
+    **GCS_RESOURCE_TREE,
+    "latest_cloud_registry_entries_file_blobs": gcs_directory_blobs.configured(
+        {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": METADATA_FOLDER, "match_regex": f".*latest/cloud.json$"}
+    ),
+    "latest_oss_registry_entries_file_blobs": gcs_directory_blobs.configured(
+        {"gcs_bucket": {"env": "METADATA_BUCKET"}, "prefix": METADATA_FOLDER, "match_regex": f".*latest/oss.json$"}
+    ),
+}
+
+CONNECTOR_TEST_REPORT_RESOURCE_TREE = {
+    **SLACK_RESOURCE_TREE,
+    **GITHUB_RESOURCE_TREE,
+    **GCS_RESOURCE_TREE,
     "latest_nightly_complete_file_blobs": gcs_directory_blobs.configured(
         {"gcs_bucket": {"env": "CI_REPORT_BUCKET"}, "prefix": NIGHTLY_FOLDER, "match_regex": f".*{NIGHTLY_COMPLETE_REPORT_FILE_NAME}$"}
     ),
@@ -102,46 +144,54 @@ RESOURCES = {
     ),
 }
 
+RESOURCES = {
+    **METADATA_RESOURCE_TREE,
+    **REGISTRY_RESOURCE_TREE,
+    **REGISTRY_ENTRY_RESOURCE_TREE,
+    **CONNECTOR_TEST_REPORT_RESOURCE_TREE,
+}
+
 SENSORS = [
     registry_updated_sensor(job=generate_registry_reports, resources_def=RESOURCES),
     new_gcs_blobs_sensor(
         job=generate_oss_registry,
-        resources_def=RESOURCES,
+        resources_def=REGISTRY_ENTRY_RESOURCE_TREE,
         gcs_blobs_resource_key="latest_oss_registry_entries_file_blobs",
-        interval=30,
+        interval=60,
     ),
     new_gcs_blobs_sensor(
         job=generate_cloud_registry,
-        resources_def=RESOURCES,
+        resources_def=REGISTRY_ENTRY_RESOURCE_TREE,
         gcs_blobs_resource_key="latest_cloud_registry_entries_file_blobs",
-        interval=30,
+        interval=60,
     ),
     new_gcs_blobs_sensor(
         job=generate_nightly_reports,
-        resources_def=RESOURCES,
+        resources_def=CONNECTOR_TEST_REPORT_RESOURCE_TREE,
         gcs_blobs_resource_key="latest_nightly_complete_file_blobs",
         interval=(1 * 60 * 60),
     ),
-    new_gcs_blobs_partition_sensor(
-        job=generate_registry_entry,
-        resources_def=RESOURCES,
-        partitions_def=registry_entry.metadata_partitions_def,
-        gcs_blobs_resource_key="all_metadata_file_blobs",
-        interval=(10 * 60 * 60),
+]
 
-    ),
-    new_gcs_blobs_partition_sensor(
-        job=generate_registry_entry,
-        resources_def=RESOURCES,
-        partitions_def=registry_entry.metadata_partitions_def,
-        gcs_blobs_resource_key="latest_metadata_file_blobs",
-        interval=60,
+SCHEDULES = [
+    ScheduleDefinition(job=add_new_metadata_partitions, cron_schedule="*/5 * * * *", tags={"dagster/priority": HIGH_QUEUE_PRIORITY}),
+    ScheduleDefinition(job=generate_connector_test_summary_reports, cron_schedule="@hourly"),
+    ScheduleDefinition(
+        cron_schedule="0 8 * * *",  # Daily at 8am US/Pacific
+        execution_timezone="US/Pacific",
+        job=generate_stale_gcs_latest_metadata_file,
     ),
 ]
 
-SCHEDULES = [ScheduleDefinition(job=generate_connector_test_summary_reports, cron_schedule="@hourly")]
-
-JOBS = [generate_registry_reports, generate_oss_registry, generate_cloud_registry, generate_registry_entry, generate_nightly_reports]
+JOBS = [
+    generate_registry_reports,
+    generate_oss_registry,
+    generate_cloud_registry,
+    generate_registry_entry,
+    generate_nightly_reports,
+    add_new_metadata_partitions,
+    generate_stale_gcs_latest_metadata_file,
+]
 
 """
 START HERE
@@ -149,6 +199,9 @@ START HERE
 This is the entry point for the orchestrator.
 It is a list of all the jobs, assets, resources, schedules, and sensors that are available to the orchestrator.
 """
+
+setup_dagster_sentry()
+
 defn = Definitions(
     jobs=JOBS,
     assets=ASSETS,