+++ airbyte-integrations/bases/debezium/src/testFixtures/java/io/airbyte/integrations/debezium/CdcSourceTest.java
@@ -770,6 +770,7 @@ public abstract class CdcSourceTest {
     // stream with PK
     streams.get(0).setSourceDefinedCursor(true);
     addCdcMetadataColumns(streams.get(0));
+    addCdcDefaultCursorField(streams.get(0));
 
     final AirbyteStream streamWithoutPK = CatalogHelpers.createAirbyteStream(
         MODELS_STREAM_NAME + "_2",
@@ -779,6 +780,7 @@ public abstract class CdcSourceTest {
         Field.of(COL_MODEL, JsonSchemaType.STRING));
     streamWithoutPK.setSourceDefinedPrimaryKey(Collections.emptyList());
     streamWithoutPK.setSupportedSyncModes(List.of(SyncMode.FULL_REFRESH));
+    addCdcDefaultCursorField(streamWithoutPK);
     addCdcMetadataColumns(streamWithoutPK);
 
     final AirbyteStream randomStream = CatalogHelpers.createAirbyteStream(
@@ -790,6 +792,8 @@ public abstract class CdcSourceTest {
         .withSourceDefinedCursor(true)
         .withSupportedSyncModes(Lists.newArrayList(SyncMode.FULL_REFRESH, SyncMode.INCREMENTAL))
         .withSourceDefinedPrimaryKey(List.of(List.of(COL_ID + "_random")));
+
+    addCdcDefaultCursorField(randomStream);
     addCdcMetadataColumns(randomStream);
 
     streams.add(streamWithoutPK);
@@ -814,6 +818,8 @@ public abstract class CdcSourceTest {
   protected abstract void removeCDCColumns(final ObjectNode data);
 
   protected abstract void addCdcMetadataColumns(final AirbyteStream stream);
+
+  protected abstract void addCdcDefaultCursorField(final AirbyteStream stream);
 
   protected abstract Source getSource();