+++ airbyte-integrations/connectors/destination-snowflake/src/test-integration/java/io/airbyte/integrations/destination/snowflake/typing_deduping/SnowflakeInternalStagingTypingDedupingTest.java
@@ -1,0 +1,8 @@
+package io.airbyte.integrations.destination.snowflake.typing_deduping;
+
+public class SnowflakeInternalStagingTypingDedupingTest extends AbstractSnowflakeTypingDedupingTest {
+  @Override
+  protected String getConfigPath() {
+    return "secrets/1s1t_internal_staging_config.json";
+  }
+}