+++ airbyte-integrations/connectors/source-mixpanel/unit_tests/test_streams.py
@@ -439,15 +439,12 @@ def test_export_stream(requests_mock, export_response, config):
 def test_export_stream_request_params(config):
     stream = Export(authenticator=MagicMock(), **config)
     stream_slice = {"start_date": "2017-01-25T00:00:00Z", "end_date": "2017-02-25T00:00:00Z"}
-    stream_state = {"date": "2021-06-16T17:00:00"}
-
-    request_params = stream.request_params(stream_state=None, stream_slice=stream_slice)
-    assert "where" not in request_params
 
     request_params = stream.request_params(stream_state={}, stream_slice=stream_slice)
     assert "where" not in request_params
 
-    request_params = stream.request_params(stream_state=stream_state, stream_slice=stream_slice)
+    stream_slice["time"] = "2021-06-16T17:00:00"
+    request_params = stream.request_params(stream_state={}, stream_slice=stream_slice)
     assert "where" in request_params
     timestamp = int(pendulum.parse("2021-06-16T17:00:00Z").timestamp())
     assert request_params.get("where") == f'properties["$time"]>=datetime({timestamp})'