+++ airbyte-integrations/connectors/source-twilio/unit_tests/test_source.py
@@ -29,9 +29,12 @@ from source_twilio.streams import (
     OutgoingCallerIds,
     Queues,
     Recordings,
+    Step,
     Transcriptions,
     UsageRecords,
     UsageTriggers,
+    UserConversations,
+    Users,
     VerifyServices,
 )
 
@@ -96,11 +99,14 @@ def test_check_connection_handles_exceptions(mocker, config, exception, expected
         (OutgoingCallerIds),
         (Queues),
         (Recordings),
+        (Step),
         (Transcriptions),
         (UsageRecords),
         (UsageTriggers),
         (Conversations),
         (ConversationParticipants),
+        (Users),
+        (UserConversations),
         (VerifyServices),
     ],
 )